// libs/vessel/src/doctypes/tile/tile-shape.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * JSON Schema for Tile file
 */
export type TileShape = TileShapeBase | WithSignature;
export type ThreeIdentifier = string;

export interface TileShapeBase {
  doctype: string;
  owners: ThreeIdentifier[];
  content: {
    [k: string]: unknown;
  };
}
export interface WithSignature {
  doctype: string;
  owners: ThreeIdentifier[];
  content: {
    [k: string]: unknown;
  };
  iss: ThreeIdentifier;
  header: {
    typ?: string;
    alg?: string;
  };
  signature: string;
  [k: string]: unknown;
}